Changed vertex color prop display
The vertex color enabled bool property is now displayed on the N key panel only when the selected geometry object actually has an use for vertex color.

For instance, collision meshes and nav meshes only use the vertex position of the mesh, thus, the property will not be displayed for those types of meshes.