Changed default vertex color value
The new default vertex color value is set to <1,1,1,0>
This is because of the way that vertex coloring works in Magicka.

The RGB elements of the vector are used to color the surface, thus using <1,1,1> will guarantee that when using a vertex color enabled effect (material), the surface will remain with the default texture's color.

The alpha channel is used to lerp between the 2 texture sets on deferred effects, and the value 0 corresponds to the first texture set, so it is ideal to have the alpha value to 0 by default.

This default value is irrelevant when using effects with the use vertex color property set to false.