Unified mesh triangulation and bm creation process
The mesh triangulation and bmesh creation process now finally makes use of the get_triangulated_mesh() function on all mesh generation functions, rather than manually handling mesh triangulation in some cases.

The get_triangulated_mesh function has been modified to now return both the mesh and bm, and the user is responsible of freeing the bm manually whenever this function is called, as the bm could be used.

TODO : Maybe add support for an input boolean parameter that could be used to determine whether the user wants the bm to be freed automatically or not?