Added check for material users on get stage
Meshes are added multiple times, once for every material they have, to be processed and exported.

If a mesh has a material that is not used within said mesh (it has no polygon that use said material), then it would be equivalent as adding an empty mesh, bypassing the empty mesh on the get stage.

This has now been fixed by checking during the mesh getting process whether a mesh has any polygons that use a given material at a certain index.

The reason we do this rather than actually implementing BiTreeNodes is that BiTreeNodes would generate greater memory fragmentation due to the fact that they are a binary tree structure.

On top of that, they do not respect contiguous memory buffers for vertex buffers and index buffers, since a new node is created for every contiguous segment of a mesh that makes use of a given material. That means that if a mesh uses 2 materials, it is not guaranteed that it will only have 2 nodes or 1 node + root mesh. It can have as many nodes as there are polygons, since BiTreeNodes can only contain contiguous polygons within its buffers. For example, if a mesh has 2 disconnected faces that make use of a secondary material, the mesh would require 2 nodes to represent both polygons, which is highly inefficient, and would make it so that the exporter would generate maps that waste more memory than required.

Considering how Magicka is a 32 bit program and there's already a 2GB memory consumption, the remaining budget is of 2GB, so it would be ideal to not waste anything so as to allow generating maps as large and complex as possible without causing problems to the user.