Implemented force field data material JSON files support
Force fields now use material JSON files just like all other geometric objects with material data.

This is because its embedded configuration is actually used under the hood 1:1 to generate an effect on Magicka's code, which means that it makes more sense to be able to edit this data through a material JSON file for ease of use and consistency, specially since it can reference texture file path strings.