Changed default color value for force field
The default force field color value is now set to <0,0,0>, as that is most likely the expected value for most players.

The reason for this is that this is the value that is used on the only instance in the base game where force fields are used, giving the expected result for anyone who wants to replicate force fields are they are in game.