Modified DataGeneratorMap to inherit from DataGenerator
The make scene data method has also been modified to make use of the generic make xnb file method.

Note that in the future it would make sense to rework this so that we could directly call make_xnb_file rather than going through make_scene_data, but still, it does not matter anyway so it's ok like this.