Refactored object properties registering to be separated by scene mode type
Each scene operation mode now has its own set of register, unregister and update object properties methods. This allows to separate the logic more easily so that we don't have all properties bundled in the same place without any form of organization.

TODO : refactor the name of the properties and move common properties such as "export" and "collision" to a generic properties method.