Added code to check if a bone has an illegal or reserved name
The names Root and RootNode are special and need to be present in all physics entity XNB files, so they are considered "reserved" / "illegal" from the point of view of MagickCow.

Also, note for the future : these bones ALWAYS have as tranform the identity matrix, so that means they are always located at <0,0,0>, have a scale of <1,1,1> and no rotation.