Implemented new scene root based scene rotation method
This new rotate scene method is based around the idea that the new scene editing system will be composed by scenes with a root object, so we only rotate the root, which prevents us from having to deal with a parentless empty fix step.

It also helps reduce the amount of floating point corrections that would need to take place when performing rotations, since the only objects with potentially lost rotation precission would be the root object.