Added test Y up converstion through change of base matrix
NOTE : This conversion works great but now vertex operations require multiple calculations. This could incur in a performance penalty with how slow these operations are in Python, so it would make sense to roll this back.

The good news is that the world rotation is no longer required to get the correct results.

The bad news is that a lot of the code needs to be modified to work with the new system.