Added scene saving and restoration code
The scene now does not require performing temporary copies to apply modifiers and extract mesh data and risking accidentally leaking memory by not freeing them. It also does not require undoing scene rotations or anything like that. All of this is now handled by the scene save enforcement system.

The exporter requires the scene to be saved before exporting it.
After the export process is performed, the scene is restored to its prior state.

This allows greater simplification when working with code that requires applying destructive modifiers to get the desired results.

TODO : Put the restoration code into a finally block so that the system is fully exception proof and always restores the scene to a fully functional point where no accidental changes have been applied.