Added origin file name writing in mcow script generator
The final generated script will have a comment appended before appending the contents of each file. The appended comment contains a string that describes the relative path in which the original file was contained, as well as the original filename.

This allows for easier tracking of file origins, making it easier to debug and to reconstruct the original source from the generated file in case that it is required.