Moved panel classes from Operators folder into Panels subdirectory
The panel classes are not considered "operator" classes in Blender's python API, so it doesn't really make sense to have them here.

They have been moved for greater organization and consistency, which will be useful when merging experimental features with the main code.
(btw, shouldn't I be using a fucking branch for experimental features to merge with main branch? like wtf...)