Modified materials code to receive materials parameter for bpy type object
This is an experimental sort of change. Rather than hardcoding the line material = bpy.types.Material in all props functions like for the object and scene properties, I'm passing them as a param this time so as to prevent having to rewrite the same line multiple times in the future in the event that the materials type class changes.