Added new base mcow exception type and modified all exception types to inherit from base mcow exception class
This change allows having a catch-all type of mcow exception that allows detecting any type of magick cow exception.

This easily allows still printing the entire stack trace when a base Exception that goes unhandled takes place without any extra work.

It also enables the code to be capable of handling mcow exceptions in a specific way in future implementations of mcow.