Fixed implementation of trigger import
The new import function now properly handles translating triggers to relative coordinates so that the imported triggers are located exactly where they would be in magicka.

Scale and position translation is now properly handled.

Note that even with the currently hacky implementation of the 90º rotation export, this implementation works 100% correctly, and it is compatible both with properly implemented and hackily implemented exporters, due to the fact that we use relative vectors (x_vec, y_vec and z_vec, being the forward, right and up vectors) for the translation to the in-editor / Blender position.