Re-generated main mcow file
NOTE : Maybe in the future it would be interesting to rework the way the export pipeline works so that it is a bit cleaner and easier to use, kinda like the far more light weight import pipeline?

Even tho there's a lot of legacy cruft from when I didn't really know how I was going to implement everything, and I was learning the blender python API and I was still reverse engineering a lot of the basics of the Magicka map file format... there's still plenty of stuff that is going to be hard to re-work considering the way that exporting data in Blender works... but there's still plenty of stuff that could probably be improved with a second go at it.