Added hacky temporary workaround to ensure that imported animated level parts have the right transform
The origin point transform for the animated level parts does not necessarily have to be the same transform as the one used on their first animation frame.

To allow objects to be generated with the correct and expected transform (location, rotation and scale), a temporary workaround has been added, where the transform of all animated level parts is modified to match that of the first animation frame.