Added BasePath.py
File that contains utility function to get the base path globally.

The idea is that users will be able to define both a global program-wide base path and a local scene-specific base path override, and this function should be a common interface to get this path without having to perform checks manually inline on every single call site.

TODO : Implement