Added support to apply generated UV data on generated meshes on import
The generated meshes on import now have a new UV layer created where the imported UV data is applied and stored.

Note that the code is currently implemented with per-vertex UV loading because the input mesh format for the importer stores the UVs per vertex, so the loading code is far simpler and more performant than if we had to do it Blender's way.