Added alpha masking support
NOTE : It looks ugly for some non-alpha masked textures that do have an alpha channel. Not sure how Magicka's render engine handles these exactly, but it's probably using the alpha channel as a way to lerp between the first and second texture sets instead, so this should be fixed in the future.