Fixed typo in exporter code
The key name for the flag that determines whether an animated level part affects shields or not used to be affectsShields but at some point it was changed on mpup to affectShields. This change was not made on this side of the code, which caused this error.

This issue has now been fixed by correcting the code to use the correct dictionary key.