Modified export pipeline to enforce object mode during export
Export mode is required for many of the destructive ops to work (applying modifiers, etc...).

To prevent the addon from throwing errors during the export process, we change the current mode to object mode, which allows the export process to work nicely.

The user will return to whatever mode they were when the blend file was saved.