Added effect deferred node group blend file
This blend file contains a simple scene that is responsible for storing the node group for the deferred effect materials.

Using this node group and adding it to the scene through the blender scene appending utility functions is far simpler than generating the entire blender node tree through python code, which is why this decision has been made.

Also do note that some absolute paths are found on the test texture assets, but those dont really matter because they are there just for testing and will never make it into the user's blender scenes. Ofc the paths don't contain any sensitive data so it doesn't matter much to me either.