Modified texture node creation so that no nodes are created if the texture data could not be loaded
If the specified texture path is absent on the input json effect data, then the generated material would have rendering issues if an empty texture node was added. To solve this issue, no node is created when the texture path string is empty or if the data could not be loaded.